oligostore/requirements.txt
Django==5.0
primer3-py
pydna
pandas
numpy
matplotlib
biopython
gunicorn
psycopg[binary]